+++ src/__mocks__/response/realEvents.json
@@ -1,1 +1,1 @@
-{"events":[{"id":"aa91a73c-f76f-4dcb-905d-4db936da562d","title":"ㄹㄴㅇㄹㄴㅇㅁㄹ123123","date":"2025-08-21","startTime":"04:29","endTime":"05:29","description":"","location":"","category":"업무","repeat":{"type":"none","interval":1},"notificationTime":10},{"id":"9b26e016-d517-48c1-8108-aa4cf5354756","title":"새로운 회의","date":"2025-08-21","startTime":"21:30","endTime":"23:30","description":"회의 설명","location":"회의실 A","category":"기타","repeat":{"type":"none","interval":1},"notificationTime":10}]}
+{"events":[{"id":"9b26e016-d517-48c1-8108-aa4cf5354756","title":"새로운 회의","date":"2025-08-21","startTime":"21:30","endTime":"23:30","description":"회의 설명","location":"회의실 A","category":"기타","repeat":{"type":"none","interval":1},"notificationTime":10}]}
